requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.6.0
pyyaml==6.0.1

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# ML/AI frameworks
torch==2.1.2
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
transformers==4.37.0

# Anthropic Claude API
anthropic==0.18.0

# Data validation and processing
jsonschema==4.21.1
marshmallow==3.20.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==24.1.0

# Task queue
celery==5.3.6
kombu==5.3.5

# Utilities
click==8.1.7
tqdm==4.66.1
requests==2.31.0
aiohttp==3.9.3

# Development tools
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.6
